requirements.txt
Flask==3.0.3
Werkzeug==3.0.2
spacy==3.7.4
pdfplumber==0.11.0
python-docx==1.1.2
gunicorn==22.0.0
